+++ src/main/java/com/task/taskmanager/infrastructure/service/impl/StateServiceImpl.java
@@ -26,8 +26,8 @@ public class StateServiceImpl implements StateService{
 	}
 	
 	@Override
-	public State getStateById(Integer idState) {
-		State state = stateRepository.findById(idState).orElse(null);
+	public State getStateById(Long idState) {
+		State state = stateRepository.findById(idState.intValue()).orElse(null);
         return state;
 	}
 
@@ -37,9 +37,9 @@ public class StateServiceImpl implements StateService{
 	}
 
 	@Override
-	public boolean deleteState(Integer idState) {
+	public boolean deleteState(Long idState) {
 		try {
-			stateRepository.deleteById(idState);
+			stateRepository.deleteById(idState.intValue());
             return true;
         } catch (Exception e) {
             return false;
