+++ src/main/java/com/task/taskmanager/infrastructure/service/spec/StateService.java
@@ -8,8 +8,8 @@ public interface StateService {
 
 	public List<State> getAllState();
 	public State createState(State state);
-	public State getStateById(Integer idState);
+	public State getStateById(Long idState);
 	public State updateState(State state);
-	public boolean deleteState(Integer idState);
+	public boolean deleteState(Long idState);
 	
 }
